Implement Decode

+++ Huffman/Bytes/DictionarySerializerResult.cs
@@ -5,6 +5,6 @@ namespace Huffman.Bytes
     public class DictionarySerializerResult
     {
         public Dictionary<byte, long> Dictionary { get; set; }
-        public long SizeOfBytes { get; set; }
+        public int SizeOfBytes { get; set; }
     }
 }
